refactor(studio): F0 foundation for GraphQL-first rebuild
Shared contract that unblocks the parallel data-layer migration (Linear EBE-95):
- client.ts: graphql() gains an endpoint param + adminGraphql() for Karrio's
  second /admin/graphql schema (users/worker_health/configs/usage), keeping the
  401->refresh retry.
- hooks/_shared.ts: keyExtra + graphqlEdges(endpoint) helpers.
- Split hooks.ts into file-disjoint domain modules so units A-E never collide:
  hooks/{ship,build,govern,resources}.ts; hooks.ts is now a thin barrel.
- STUDIO_GRAPHQL_REBUILD.md: in-repo contract with the LIVE-verified tenant +
  admin GraphQL field lists, per-hook source mapping, and acceptance criteria.

Behaviour-preserving refactor (no endpoint changes yet). tsc clean, vite build
clean, mocked studio suite passing. Units A-E rewrite each domain file to the
canonical GraphQL sources next.
